# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Auth service API documentation",
        "title": "api gat way",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/game/complete_level": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Complete an ongoing game level",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Complete Level",
                "parameters": [
                    {
                        "description": "Complete level request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/game.CompleteLevelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Level completed successfully",
                        "schema": {
                            "$ref": "#/definitions/game.CompleteLevelResponse"
                        }
                    },
                    "500": {
                        "description": "Error while completing level",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/game/get_achievements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve achievements based on criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Get Achievements",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player ID",
                        "name": "player_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Achievements list",
                        "schema": {
                            "$ref": "#/definitions/game.AchievementsResponse"
                        }
                    },
                    "400": {
                        "description": "Missing required query parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while fetching achievements",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/game/get_challenge": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve details of a challenge by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Get Challenge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Challenge ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Challenge details",
                        "schema": {
                            "$ref": "#/definitions/game.Challenge"
                        }
                    },
                    "400": {
                        "description": "Missing required query parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while fetching challenge",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/game/get_leaderboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve leaderboard based on criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Get Leaderboard",
                "responses": {
                    "200": {
                        "description": "Leaderboard details",
                        "schema": {
                            "$ref": "#/definitions/game.LeaderboardResponse"
                        }
                    },
                    "400": {
                        "description": "Missing required query parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while fetching leaderboard",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/game/get_levels": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of the game levels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Get Levels",
                "responses": {
                    "200": {
                        "description": "Game levels details",
                        "schema": {
                            "$ref": "#/definitions/game.GetLevelsResponse"
                        }
                    },
                    "400": {
                        "description": "Missing required query parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while fetching levels",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/game/start_level": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Start a new game level",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Start Level",
                "parameters": [
                    {
                        "description": "Start level request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/game.StartLevelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Level started successfully",
                        "schema": {
                            "$ref": "#/definitions/game.StartLevelResponse"
                        }
                    },
                    "500": {
                        "description": "Error while starting level",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/game/submit_challenge": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit a response to a challenge",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Submit Challenge",
                "parameters": [
                    {
                        "description": "Submit challenge request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/game.SubmitChallengeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Challenge submitted successfully",
                        "schema": {
                            "$ref": "#/definitions/game.SubmitChallengeResponse"
                        }
                    },
                    "500": {
                        "description": "Error while submitting challenge",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "game.Achievement": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "earned_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "game.AchievementsResponse": {
            "type": "object",
            "properties": {
                "achievements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/game.Achievement"
                    }
                }
            }
        },
        "game.Answer": {
            "type": "object",
            "properties": {
                "question_id": {
                    "type": "string"
                },
                "selected_option": {
                    "type": "integer"
                }
            }
        },
        "game.Challenge": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/game.Question"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "game.CompleteLevelRequest": {
            "type": "object",
            "properties": {
                "level_id": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                },
                "xpearned": {
                    "type": "integer"
                }
            }
        },
        "game.CompleteLevelResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "new_level_unlocked": {
                    "type": "string"
                },
                "xp_earned": {
                    "type": "integer"
                }
            }
        },
        "game.GetLevelsResponse": {
            "type": "object",
            "properties": {
                "levels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/game.Level"
                    }
                }
            }
        },
        "game.LeaderboardEntry": {
            "type": "object",
            "properties": {
                "level_id": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "user_xp": {
                    "type": "integer"
                }
            }
        },
        "game.LeaderboardResponse": {
            "type": "object",
            "properties": {
                "leaderboard": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/game.LeaderboardEntry"
                    }
                }
            }
        },
        "game.Level": {
            "type": "object",
            "properties": {
                "challenges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/game.Challenge"
                    }
                },
                "cid": {
                    "type": "integer"
                },
                "completed": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "levelid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "required_xp": {
                    "type": "integer"
                }
            }
        },
        "game.Question": {
            "type": "object",
            "properties": {
                "correct_option": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "game.StartLevelRequest": {
            "type": "object",
            "properties": {
                "level_id": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
        "game.StartLevelResponse": {
            "type": "object",
            "properties": {
                "first_challenge_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "game.SubmitChallengeRequest": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/game.Answer"
                    }
                },
                "challenge_id": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
        "game.SubmitChallengeResponse": {
            "type": "object",
            "properties": {
                "correct_answers": {
                    "type": "integer"
                },
                "feedback": {
                    "type": "string"
                },
                "total_questions": {
                    "type": "integer"
                },
                "xp_earned": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}